+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [2.1.0]
+### Added
+- support for custom styles for title, description, note texts, comments
+- custom star color
+- "Later" button
+
+### Changed
+- **Breaking change:** Added "onNeutralButtonClicked" method to listener
+
+[2.1.0]: https://github.com/stepstone-tech/android-material-app-rating/compare/v2.0.0...v2.1.0
+
+
+
 ## [2.0.0]
 ### Added
 - support for custom hint text and color in comment box